refactor: 이메일 인증 기능을 command에서 query로 이동 (#25)
