+++ src/main/java/com/harusari/chainware/member/command/application/service/MemberCommandService.java
@@ -4,12 +4,9 @@ import com.harusari.chainware.member.command.application.dto.request.MemberCreat
 import com.harusari.chainware.member.command.application.dto.request.franchise.MemberWithFranchiseRequest;
 import com.harusari.chainware.member.command.application.dto.request.vendor.MemberWithVendorRequest;
 import com.harusari.chainware.member.command.application.dto.request.warehouse.MemberWithWarehouseRequest;
-import com.harusari.chainware.member.command.application.dto.response.EmailExistsResponse;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface MemberCommandService {
-
-    EmailExistsResponse checkEmailDuplicate(String email);
 
     void registerHeadquartersMember(MemberCreateRequest memberCreateRequest);
