--- src/main/java/com/harusari/chainware/member/query/mapper/MemberMapper.java
@@ -1,10 +1,0 @@
-package com.harusari.chainware.member.query.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface MemberMapper {
-
-
-
-}
