+++ src/main/java/com/harusari/chainware/member/query/repository/MemberQueryRepository.java
@@ -1,0 +1,10 @@
+package com.harusari.chainware.member.query.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface MemberQueryRepository {
+
+    boolean existsByEmail(String email);
+
+}
