+++ src/main/java/com/harusari/chainware/member/query/service/MemberQueryService.java
@@ -1,7 +1,9 @@
 package com.harusari.chainware.member.query.service;
 
+import com.harusari.chainware.member.command.application.dto.response.EmailExistsResponse;
+
 public interface MemberQueryService {
 
-
+    EmailExistsResponse checkEmailDuplicate(String email);
 
 }
